pyproject.toml
[project]
name = "llm-scratch"
version = "0.1.0"
description = "Scratch experiments for a decoder-only autoregressive Transformer and BPE tokenizer."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "hydra-core>=1.3",
  "loguru>=0.7",
  "numpy>=1.26",
  "torch>=2.2",
  "tqdm>=4.66",
  "wandb>=0.19",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
]
notebook = [
  "ipykernel>=6.29",
  "matplotlib>=3.9",
]

[tool.uv]
default-groups = ["dev", "notebook"]
